requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
numpy
pillow
mediapipe
python-multipart
huggingface_hub